feat(react-ui): rename toast folder
