+++ libs/react-ui/src/index.ts
@@ -4,7 +4,7 @@ export {
   ButtonSize,
 } from './lib/button/button';
 
-export { default as Toast } from './lib/Toast-rename';
+export { default as Toast } from './lib/toast';
 
 export * from './lib/context';
 export * from './lib/hooks';
